Started the Scaffold for Enemy Hits.

+++ Source/timbermvp/Public/Character/TimberCharacterBase.h
@@ -36,6 +36,6 @@ public:
 	float MaxBackwardSpeed = 500.f;
 
 	UFUNCTION()
-	void DecreaseHealth(int DecreaseAmount);
+	void DecreaseHealth(float DecreaseAmount);
 	
 };
